Merge remote-tracking branch 'webkit/master' into master.

+++ Source/WebCore/platform/graphics/win/GraphicsContextCGWin.cpp
@@ -10,10 +10,10 @@
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
  *
- * THIS SOFTWARE IS PROVIDED BY APPLE COMPUTER, INC. ``AS IS'' AND ANY
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. ``AS IS'' AND ANY
  * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL APPLE COMPUTER, INC. OR
+ * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL APPLE INC. OR
  * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
  * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
  * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
@@ -228,10 +228,10 @@ void GraphicsContext::drawLineForDocumentMarker(const FloatPoint& point, float w
     
     // Dash lengths for the top and bottom of the error underline are the same.
     // These are magic.
-    static const float edge_dash_lengths[] = {2.0f, 2.0f};
-    static const float middle_dash_lengths[] = {2.76f, 1.24f};
-    static const float edge_offset = -(edge_dash_lengths[1] - 1.0f) / 2.0f;
-    static const float middle_offset = -(middle_dash_lengths[1] - 1.0f) / 2.0f;
+    static const CGFloat edge_dash_lengths[] = {2.0f, 2.0f};
+    static const CGFloat middle_dash_lengths[] = { 2.76f, 1.24f };
+    static const CGFloat edge_offset = -(edge_dash_lengths[1] - 1.0f) / 2.0f;
+    static const CGFloat middle_offset = -(middle_dash_lengths[1] - 1.0f) / 2.0f;
 
     // Line opacities.  Once again, these are magic.
     const float upperOpacity = 0.33f;
